First try to independently plot subpanels
Can be called with an extra argument like so
gridspeccer fig_setup.py --mplrc ../gridspeccer/defaults/tex_matplotlibrc --independent-plots-filetype .svg
It is set up in the `get_gridspec` function, one can give a dictionary
of filenames, gridspecs that should be covered plus an additional
Extension to the area of the gridspec for flexibility.
The latter is needed as gridspecs by default do not include the axes for
some reason.